# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{id}/follow": {
      "post": {
        "operationId": "UsersController_followById",
        "summary": "Follow user by ID",
        "description": "Current user follows the user with the given ID. Returns 204 on success.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Following"
          },
          "400": {
            "description": "Cannot follow yourself",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_unfollowById",
        "summary": "Unfollow user by ID",
        "description": "Current user stops following the user with the given ID. Returns 204 on success.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Unfollowed"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users/{username}/owned": {
      "get": {
        "operationId": "UsersController_getPublicOwned",
        "summary": "Get user collection",
        "description": "Returns the viewed user’s owned games (collection) by username. Public endpoint.",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "description": "Username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owned games",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesListResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{username}/wishlist": {
      "get": {
        "operationId": "UsersController_getPublicWishlist",
        "summary": "Get user wanted list",
        "description": "Returns the viewed user’s wishlist (wanted) games by username. Public endpoint.",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "description": "Username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wishlist games",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesListResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{username}": {
      "get": {
        "operationId": "UsersController_getPublicProfile",
        "summary": "Get public profile",
        "description": "Returns public profile for the given username (id, username, followersCount, followingCount).",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "description": "Username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    }
  },
  "info": {
    "title": "BG Hub API",
    "description": "Board game collection, wishlist, play logs and social features.",
    "version": "0.0.2",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Game UUID",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Game name",
            "example": "Catan"
          },
          "externalId": {
            "type": "string",
            "description": "External provider ID (e.g. BGG)",
            "example": "13"
          },
          "apiRef": {
            "type": "string",
            "description": "External API reference",
            "example": "bgg"
          },
          "imageUrl": {
            "type": "object",
            "description": "Cover image URL",
            "nullable": true
          },
          "year": {
            "type": "object",
            "description": "Year published",
            "nullable": true,
            "example": 1995
          },
          "minPlayers": {
            "type": "object",
            "description": "Minimum number of players",
            "nullable": true,
            "example": 3
          },
          "maxPlayers": {
            "type": "object",
            "description": "Maximum number of players",
            "nullable": true,
            "example": 4
          },
          "playTime": {
            "type": "object",
            "description": "Playing time in minutes",
            "nullable": true,
            "example": 90
          },
          "complexityWeight": {
            "type": "object",
            "description": "Complexity/weight (e.g. BGG)",
            "nullable": true
          },
          "categories": {
            "description": "Category names",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mechanics": {
            "description": "Mechanic names",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "object",
            "description": "Game description",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "externalId",
          "apiRef",
          "imageUrl",
          "year",
          "minPlayers",
          "maxPlayers",
          "playTime",
          "complexityWeight",
          "categories",
          "mechanics",
          "description"
        ]
      },
...
      "GamesListResponseDto": {
        "type": "object",
        "properties": {
          "games": {
            "description": "List of games",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameDto"
            }
          }
        },
        "required": [
          "games"
        ]
      },
...
      "PublicProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User UUID",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "johndoe"
          },
          "followersCount": {
            "type": "number",
            "description": "Number of followers",
            "example": 42
          },
          "followingCount": {
            "type": "number",
            "description": "Number of users this user follows",
            "example": 10
          },
          "displayName": {
            "type": "string",
            "description": "Display name",
            "example": "John"
          },
          "bio": {
            "type": "string",
            "description": "Short bio",
            "example": "Board game enthusiast."
          },
          "avatarUrl": {
            "type": "string",
            "description": "Avatar image URL",
            "example": "https://example.com/avatar.jpg"
          }
        },
        "required": [
          "id",
          "username",
          "followersCount",
          "followingCount"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code",
            "example": 404
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Resource not found"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}